requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
cohere==4.0.0
google-cloud==0.34.0
python-dotenv==1.0.0
PyMuPDF==1.18.19
python-docx==0.8.11
s